backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "my-backend"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlmodel",
    "python-dotenv",
    "aiosqlite>=0.22.1",
    "greenlet>=3.3.0",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "cryptography>=46.0.3",
    "bcrypt>=5.0.0",
    "httpx>=0.28.1",
    "python-multipart>=0.0.21",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["main.py"]

